refactor: split pull_request_sync.py into focused sub-modules
Extract PR sync logic from the 995-line monolithic pull_request_sync.py
into three smaller, context-optimized modules following CLAUDE.md guidance:

- pr_asana_helpers.py: Asana API helpers (cached task/field lookup, task
  create/update/tag/comment)
- pr_processor.py: individual PR and reviewer processing logic
- pr_sync_core.py: top-level orchestration (sync_pull_requests,
  process_repository_pull_requests, process_closed_pull_requests)

pull_request_sync.py is retained as a re-export facade for backward
compatibility. All tests updated to patch the correct module namespaces.
Documentation updated across AGENTS.md, CLAUDE.md, copilot-instructions.md,
and TESTING_EXAMPLES.md.
